Integer 直接比较 存在bug 需要用 intValue() 转成值比较

+++ LeetCode/src/com/eco/study/leetcode/string/Lesson383.java
@@ -62,7 +62,7 @@ class Solution383 {
         Map<String,Integer> r = countCharSumToMap(ransomNote);
         Map<String,Integer> m = countCharSumToMap(magazine);
         for (String k : r.keySet()) {
-            if(m.getOrDefault(k,0) < r.get(k)){
+            if(m.getOrDefault(k,0).intValue() < r.get(k).intValue()){
                 return false;
             }
         }
